+++ trello-clone/src/main/resources/db/migration/V17__add_checkable_items_table.sql
@@ -2,8 +2,8 @@ create table checkable_items
 (
     id           uuid primary key,
     description  varchar(50),
-    checked      bool     not null default false,
+    checked      bool not null default false,
 
-    checklist_id uuid     not null references checklists (id),
-    position     smallint not null
+    checklist_id uuid not null references checklists (id),
+    position     int  not null
 )
